add in cast message, tweaks to pathfinding, start breaking out control messages into individual partial classes to clean up world.cs
